refactor: 서버 컴포넌트를 위해 page 분리
