+++ src/app/Providers.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import StyledComponentsRegistry from '../styles/registry';
+import StyledComponentsRegistry from './registry';
 import { ThemeProvider } from 'styled-components';
 import theme from '../styles/theme';
 import GlobalStyles from '../GlobalStyles';
@@ -8,7 +8,10 @@ import GlobalStyles from '../GlobalStyles';
 const Providers = (props: React.PropsWithChildren) => {
   return (
     <StyledComponentsRegistry>
-      <ThemeProvider theme={theme}><GlobalStyles />{props.children}</ThemeProvider>
+      <ThemeProvider theme={theme}>
+        <GlobalStyles />
+        {props.children}
+      </ThemeProvider>
     </StyledComponentsRegistry>
   );
 };
