+++ src/app/layout.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import Providers from './Providers';
 
 export default function RootLayout({ children }: { children: React.ReactNode }) {
