+++ src/app/registry.tsx
